+++ terraform/aws/dev/s3.tf
@@ -6,6 +6,12 @@ module "installers" {
   bucket = join(var.delimiter, [var.name, var.stage, "installers", random_string.root.id])
   acl    = "private"
 
+  # todo enable bucket object storage tiering
+  # intelligent_tiering = {
+  #   access_tier = "ARCHIVE_ACCESS"
+  #   days        = 30
+  # }
+
   server_side_encryption_configuration = {
     rule = {
       kms_master_key_id = aws_kms_alias.installers.arn
@@ -20,6 +26,27 @@ module "installers" {
 
 resource "aws_s3_object" "text_shellscript" {
   for_each = fileset("${path.module}/installers", "*.sh")
+
+  lifecycle {
+    prevent_destroy = true
+  }
+
+  content_type           = "text/x-shellscript"
+  server_side_encryption = "AES256"
+
+  etag   = filemd5("${path.module}/installers/${each.value}")
+  source = "${path.module}/installers/${each.value}"
+  key    = each.value
+
+  bucket = module.installers.s3_bucket_id
+}
+
+resource "aws_s3_object" "title_archives" {
+  for_each = fileset("${path.module}/installers", "{*.zip||*.tar.bz2||*.tar.xz}")
+
+  lifecycle {
+    prevent_destroy = true
+  }
 
   content_type           = "text/x-shellscript"
   server_side_encryption = "AES256"